+++ examples/train-and-predict.py
@@ -1,6 +1,10 @@
 """
-Using the same pipeline for fitting and predictions
+Traind and predict pipeline
+===========================
+
+Example showing how to build a training and a prediction pipeline
 """
+import tempfile
 from pathlib import Path
 
 from ploomber import DAG
@@ -37,6 +41,8 @@ def _features(upstream, product):
 
 
 def _join_features(upstream, product):
+    """Join raw data with generated features
+    """
     a = pd.read_parquet(str(upstream['get']))
     b = pd.read_parquet(str(upstream['features']))
     df = a.join(b)
@@ -44,7 +50,7 @@ def _join_features(upstream, product):
 
 
 def _fit(upstream, product):
-    """Fit preprocessor and model
+    """Fit model and generate classification report
     """
     df = pd.read_parquet(str(upstream['join']))
     X = df.drop('target', axis='columns')
@@ -74,7 +80,9 @@ def _new_obs(product, values):
     df.to_parquet(str(product))
 
 
-def _score(product, upstream, model):
+def _pred(product, upstream, model):
+    """De-serialize model and make a new prediction
+    """
     clf = joblib.load(model)
     df = pd.read_parquet(str(upstream['join']))
     df = df.drop('target', axis='columns')
@@ -83,35 +91,65 @@ def _score(product, upstream, model):
     score.to_csv(str(product), index=False)
 
 
-def make_pipeline(dag, get):
+def add_fts(dag):
+    """Modify a pipeline with a "get" task to generate features and join them
+    """
     fts = PythonCallable(_features, File('features.parquet'), dag,
                          name='features')
     join = PythonCallable(_join_features, File('join.parquet'), dag,
                           name='join')
 
-    get >> fts
+    dag['get'] >> fts
 
-    (get + fts) >> join
+    (dag['get'] + fts) >> join
 
     return dag
 
 
-dag_fit = DAG()
-get = PythonCallable(_get, File('data.parquet'), dag_fit, name='get')
-make_pipeline(dag_fit, get)
-fit = PythonCallable(_fit, {'report': File('report.txt'),
-                            'model': File('model.joblib')}, dag_fit,
-                     name='fit')
+tmp_dir = Path(tempfile.mkdtemp())
 
+
+# build training pipeline
+dag_fit = DAG()
+get = PythonCallable(_get,
+                     File(tmp_dir / 'data.parquet'),
+                     dag_fit,
+                     name='get')
+dag_fit = add_fts(dag_fit)
+fit = PythonCallable(_fit, {'report': File(tmp_dir / 'report.txt'),
+                            'model': File(tmp_dir / 'model.joblib')},
+                     dag_fit,
+                     name='fit')
 dag_fit['join'] >> fit
 
 
-dag_pred = DAG()
-get = PythonCallable(_new_obs, File('obs.parquet'), dag_pred, name='get',
-                     params={'values': [1, 0, 10, 2]})
-make_pipeline(dag_pred, get)
-score = PythonCallable(_score, File('pred.txt'), dag_pred,
-                       name='pred',
-                       params={'model': 'model.joblib'})
+###############################################################################
+# Fit pipeline plot
+dag_fit.plot(output='matplotlib')
 
-dag_pred['join'] >> score
+
+dag_fit.build()
+
+# build prediction pipeline - pass a new observation with values [1, 0, 10, 2]
+dag_pred = DAG()
+get = PythonCallable(_new_obs,
+                     File(tmp_dir / 'obs.parquet'),
+                     dag_pred,
+                     name='get',
+                     params={'values': [1, 0, 10, 2]})
+
+dag_pred = add_fts(dag_pred)
+pred = PythonCallable(_pred, File(tmp_dir / 'pred.csv'), dag_pred,
+                      name='pred',
+                      params={'model': tmp_dir / 'model.joblib'})
+
+dag_pred['join'] >> pred
+
+###############################################################################
+# Prediction pipeline plot
+dag_pred.plot(output='matplotlib')
+
+dag_pred.build()
+
+# get prediction
+pd.read_csv(str(dag_pred['pred']))