+++ src/ploomber/clients/__init__.py
@@ -1,6 +1,6 @@
 from ploomber.clients.Client import Client
 from ploomber.clients.db import (DBAPIClient, SQLAlchemyClient,
-                                         DrillClient)
+                                 DrillClient)
 from ploomber.clients.shell import ShellClient, RemoteShellClient
 
 __all__ = ['Client', 'DBAPIClient', 'SQLAlchemyClient', 'DrillClient',