+++ src/ploomber/tasks/sql.py
@@ -5,18 +5,39 @@ from ploomber.exceptions import SourceInitializationError
 from ploomber.tasks.Task import Task
 from ploomber.sources import (SQLScriptSource,
                               SQLQuerySource,
-                              GenericSource)
+                              GenericSource,
+                              FileSource)
 from ploomber.products import File, PostgresRelation, SQLiteRelation
 from ploomber import io
+from ploomber.util import requires
 
 import pandas as pd
 
 
 class SQLScript(Task):
-    """
-    A tasks represented by a SQL script run agains a database this Task
-    does not make any assumptions about the underlying SQL engine, it should
-    work witn all DBs supported by SQLAlchemy
+    """Execute a script in a SQL database
+
+    Parameters
+    ----------
+    source: str or pathlib.Path
+        SQL script source, if str, the content is interpreted as the actual
+        script, if pathlib.Path, the content of the file is loaded
+    product: ploomber.products.Product
+        Product generated upon successful execution
+    dag: ploomber.DAG
+        A DAG to add this task to
+    name: str
+        A str to indentify this task. Should not already exist in the dag
+    client: ploomber.clients.DBAPIClient or SQLAlchemyClient, optional
+        The client used to connect to the database. Only required
+        if no dag-level client has been declared using dag.clients[class]
+    params: dict, optional
+        Parameters to pass to the script, by default, the callable will
+        be executed with a "product" (which will contain the product object).
+        It will also include a "upstream" parameter if the task has upstream
+        dependencies along with any parameters declared here. The source
+        code is converted to a jinja2.Template for passing parameters,
+        refer to jinja2 documentation for details
     """
     PRODUCT_CLASSES_ALLOWED = (PostgresRelation, SQLiteRelation)
 
@@ -39,43 +60,55 @@ class SQLScript(Task):
 
 
 class SQLDump(Task):
-    """
-    Dumps data from a SQL SELECT statement to parquet files (one per chunk)
+    """Dumps data from a SQL SELECT statement to a file(s)
 
     Parameters
     ----------
-    source: str
-        The SQL query to run in the database
-    product: File
-        The directory location for the output parquet files
-    dag: DAG
-        The DAG for this task
+    source: str or pathlib.Path
+        SQL script source, if str, the content is interpreted as the actual
+        script, if pathlib.Path, the content of the file is loaded
+    product: ploomber.products.Product
+        Product generated upon successful execution
+    dag: ploomber.DAG
+        A DAG to add this task to
     name: str
-        Name for this task
+        A str to indentify this task. Should not already exist in the dag
+    client: ploomber.clients.DBAPIClient or SQLAlchemyClient, optional
+        The client used to connect to the database. Only required
+        if no dag-level client has been declared using dag.clients[class]
     params: dict, optional
-        Extra parameters for the task
+        Parameters to pass to the script, by default, the callable will
+        be executed with a "product" (which will contain the product object).
+        It will also include a "upstream" parameter if the task has upstream
+        dependencies along with any parameters declared here. The source
+        code is converted to a jinja2.Template for passing parameters,
+        refer to jinja2 documentation for details
     chunksize: int, optional
-        Size of each chunk, one parquet file will be generated per chunk. If
-        None, only one file is created
-
+        Number of rows per file, otherwise download the entire dataset in
+        a single one. If set, the product will be a folder
+    io_handler: ploomber.io.CSVIO or ploomber.io.ParquetIO, optional
+        io handler to use (which controls the output format), currently
+        only csv and parquet are supported. Defaults to ploomber.io.CSVIO
 
     Notes
     -----
-    The chunksize parameter is set in cursor.arraysize object, this parameter
-    can greatly speed up the dump for some databases when the driver uses
-    cursors.arraysize as the number of rows to fetch on a single call
+    The chunksize parameter is also set in cursor.arraysize object, this
+    parameter can greatly speed up the dump for some databases when the
+    driver uses cursors.arraysize as the number of rows to fetch on a single
+    network trip, but this is driver-dependent, not all drivers implement
+    this (cx_Oracle does it)
     """
     PRODUCT_CLASSES_ALLOWED = (File, )
 
     def __init__(self, source, product, dag, name, client=None,
                  params=None,
-                 chunksize=10000, io_handler=None):
+                 chunksize=10000, io_handler=io.CSVIO):
         params = params or {}
         super().__init__(source, product, dag, name, params)
 
         self.client = client or self.dag.clients.get(type(self))
         self.chunksize = chunksize
-        self.io_handler = io_handler or io.CSVIO
+        self.io_handler = io_handler
 
         if self.client is None:
             raise ValueError('{} must be initialized with a client'
@@ -125,10 +158,45 @@ class SQLDump(Task):
 
 
 class SQLTransfer(Task):
-    """Transfers data from a SQL statement to a SQL relation
+    """
+    Transfers data from a SQL database to another (Note: this relies on
+    pandas, only use it for small to medium size datasets)
+
+    Parameters
+    ----------
+    source: str or pathlib.Path
+        SQL script source, if str, the content is interpreted as the actual
+        script, if pathlib.Path, the content of the file is loaded
+    product: ploomber.products.Product
+        Product generated upon successful execution. For SQLTransfer, usually
+        product.client != task.client. task.client represents the data source
+        while product.client represents the data destination.
+    dag: ploomber.DAG
+        A DAG to add this task to
+    name: str
+        A str to indentify this task. Should not already exist in the dag
+    client: ploomber.clients.SQLAlchemyClient, optional
+        The client used to connect to the database. Only required
+        if no dag-level client has been declared using dag.clients[class]
+    params: dict, optional
+        Parameters to pass to the script, by default, the callable will
+        be executed with a "product" (which will contain the product object).
+        It will also include a "upstream" parameter if the task has upstream
+        dependencies along with any parameters declared here. The source
+        code is converted to a jinja2.Template for passing parameters,
+        refer to jinja2 documentation for details
+    chunksize: int, optional
+        Number of rows to transfer on every chunk
+
+    Notes
+    ----
+    This task is *not* intended to move large datasets, but a
+    convenience way of transfering small to medium size datasets. It relies
+    on pandas to read and write, which introduces a considerable overhead.
     """
     PRODUCT_CLASSES_ALLOWED = (PostgresRelation, SQLiteRelation)
 
+    @requires(['pandas'], 'SQLTransfer')
     def __init__(self, source, product, dag, name, client=None,
                  params=None, chunksize=10000):
         params = params or {}
@@ -143,6 +211,7 @@ class SQLTransfer(Task):
         self.chunksize = chunksize
 
     def _init_source(self, source):
+        # TODO: this shoule be a FileSource
         return SQLQuerySource(source)
 
     def run(self):
@@ -164,18 +233,48 @@ class SQLTransfer(Task):
                       index=False)
 
 
+# FIXME: support other data formats apart from parquet
 class SQLUpload(Task):
-    """Upload data to a database from a parquet file
+    """
+    Upload data to a SQL database from a parquet file (Note: this relies on
+    pandas, only use it for small to medium size datasets)
 
     Parameters
     ----------
     source: str or pathlib.Path
         Path to parquet file to upload
+    product: ploomber.products.Product
+        Product generated upon successful execution. For SQLTransfer, usually
+        product.client != task.client. task.client represents the data source
+        while product.client represents the data destination.
+    dag: ploomber.DAG
+        A DAG to add this task to
+    name: str
+        A str to indentify this task. Should not already exist in the dag
+    client: ploomber.clients.SQLAlchemyClient, optional
+        The client used to connect to the database. Only required
+        if no dag-level client has been declared using dag.clients[class]
+    params: dict, optional
+        Parameters to pass to the script, by default, the callable will
+        be executed with a "product" (which will contain the product object).
+        It will also include a "upstream" parameter if the task has upstream
+        dependencies along with any parameters declared here. The source
+        code is converted to a jinja2.Template for passing parameters,
+        refer to jinja2 documentation for details
+    chunksize: int, optional
+        Number of rows to transfer on every chunk
+
+    Notes
+    -----
+    This task is *not* intended to move large datasets, but a
+    convenience way of transfering small to medium size datasets. It relies
+    on pandas to read and write, which introduces a considerable overhead.
     """
     PRODUCT_CLASSES_ALLOWED = (PostgresRelation, SQLiteRelation)
 
+    @requires(['pandas'], 'SQLUpload')
     def __init__(self, source, product, dag, name, client=None,
-                 params=None):
+                 params=None, chunksize=None):
         params = params or {}
         super().__init__(source, product, dag, name, params)
 
@@ -185,15 +284,10 @@ class SQLUpload(Task):
             raise ValueError('{} must be initialized with a connection'
                              .format(type(self).__name__))
 
+        self.chunksize = chunksize
+
     def _init_source(self, source):
-        source = GenericSource(str(source))
-
-        if source.needs_render:
-            raise SourceInitializationError('{} does not support templates as '
-                                            'source, pass a path to a file',
-                                            self.__class__)
-
-        return source
+        return FileSource(str(source))
 
     def run(self):
         product = self.params['product']
@@ -209,17 +303,27 @@ class SQLUpload(Task):
                   index=False)
 
 
-class PostgresCopy(Task):
-    """Efficiently copy data to a postgres database using COPY (better
-    alternative to SQLUpload for postgres)
+# TODO: provide more flexibility to configure the COPY statement
+class PostgresCopyFrom(Task):
+    """Efficiently copy data to a postgres database using COPY FROM (faster
+    alternative to SQLUpload for postgres). If using SQLAlchemy client
+    for postgres is psycopg2. Replaces the table if exists.
 
     Parameters
     ----------
     source: str or pathlib.Path
         Path to parquet file to upload
+
+
+    Notes
+    -----
+    Although this task does not depend on pandas for data i/o, it still
+    needs it to dynamically create the table, after the table is created
+    the COPY statement is used to upload the data
     """
     PRODUCT_CLASSES_ALLOWED = (PostgresRelation,)
 
+    @requires(['pandas'], 'PostgresCopyFrom')
     def __init__(self, source, product, dag, name, client=None,
                  params=None, sep='\t', null='\\N', columns=None):
         params = params or {}
@@ -236,14 +340,7 @@ class PostgresCopy(Task):
         self.columns = columns
 
     def _init_source(self, source):
-        source = GenericSource(str(source))
-
-        if source.needs_render:
-            raise SourceInitializationError('{} does not support templates as '
-                                            'source, pass a path to a file',
-                                            self.__class__)
-
-        return source
+        return FileSource(str(source))
 
     def run(self):
         product = self.params['product']