+++ src/ploomber/util/__init__.py
@@ -1,6 +1,7 @@
 from ploomber.util.util import (safe_remove, image_bytes2html,
-                                        clean_up_files, isiterable)
+                                clean_up_files, isiterable, path2fig,
+                                requires)
 from ploomber.util.param_grid import Interval, ParamGrid
 
 __all__ = ['safe_remove', 'image_bytes2html', 'clean_up_files', 'Interval',
-           'ParamGrid', 'isiterable']
+           'ParamGrid', 'isiterable', 'path2fig', 'requires']