+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+0.2
+---
+
+* Improved documentation
+* Renamed PostgresCopy to PostgresCopyFrom
+* SQLUpload and PostgresCopy have now the same API
+* A few fixes to PostgresCopy (#1, #2)
+
 0.1
 ---